add function "DisplayTime(hours, minutes)" to display hours and minutes but it overwrite whole PORTD which could lead to problems witch buttons in future

+++ Libraries/Utils/utils.c
@@ -23,7 +23,7 @@ void BlinkLed()
 	LED1_OFF;
 }
 
-uint32_t millis;
+uint64_t millis;
 void Timer0_Init()
 {
 	sei();// allow interrupts
